pyproject.toml
[tool.poetry]
name = "poc-contra-in-browser"
version = "0.1.0"
description = ""
authors = ["thelonggoodbuy <thelonggoodbuy1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
django = "4.2"
django-ninja = "^1.1.0"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
pycryptodome = "^3.20.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
